Add all EntityNotFound throws on services

+++ src/main/java/com/pharius/inventoryapp/inventoryapp/Services/EstablishmentServices/DeleteEstablishmentService.java
@@ -1,10 +1,15 @@
 package com.pharius.inventoryapp.inventoryapp.Services.EstablishmentServices;
 
+import java.util.Optional;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.pharius.inventoryapp.inventoryapp.Controllers.Command;
+import com.pharius.inventoryapp.inventoryapp.Exceptions.EntityNotFoundException;
+import com.pharius.inventoryapp.inventoryapp.Exceptions.ErrorMessages;
+import com.pharius.inventoryapp.inventoryapp.Models.EstablishmentModels.Establishment;
 import com.pharius.inventoryapp.inventoryapp.Repositories.EstablishmentRepository;
 
 @Service
@@ -19,13 +24,12 @@ public class DeleteEstablishmentService implements Command<Long, Void> {
 
     @Override
     public ResponseEntity<Void> execute(Long establishmentId) {
-      
-        establishmentRepository.deleteById(establishmentId);
-
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build(); 
-
-
-
+        Optional<Establishment> foundedEstablishment = establishmentRepository.findById(establishmentId);
+        if(foundedEstablishment.isPresent()){
+            establishmentRepository.deleteById(establishmentId);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+        throw new EntityNotFoundException(ErrorMessages.ENTITY_NOT_FOUND, "Establishment");
     }
 
 }
